+++ lumenairy/__init__.py
@@ -1099,7 +1099,7 @@ load_zemax_prescription_txt = _deprecated_alias(
     version_removed='6.0',
 )
 
-__version__ = "5.35.3"
+__version__ = "5.35.4"
 
 #
 # __all__ is grouped by user-journey tier: